ng2: move admin.ng2module to new file
