+++ generators/client-2/templates/src/main/webapp/app/admin/_admin.module.ts
@@ -1,10 +1,5 @@
 import { upgradeAdapter } from '../upgrade_adapter';
 
-import { XSRFStrategy, CookieXSRFStrategy } from '@angular/http';
-import { FormsModule } from '@angular/forms';
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-
 import { AdminStateConfig } from './admin.state';
 import { AuditStateConfig } from './audits/audits.state';
 import { ConfigStateConfig } from './configuration/configuration.state';
@@ -30,14 +25,7 @@ import { <%=jhiPrefixCapitalized%>HealthModalController } from "./health/health-
 import { <%=jhiPrefixCapitalized%>HealthCheckComponent } from './health/health.component';
 import { <%=jhiPrefixCapitalized%>ConfigurationComponent } from './configuration/configuration.component';
 
-import { AuditsService } from './audits/audits.service';
 import { <%=jhiPrefixCapitalized%>HealthService } from './health/health.service';
-import { LogsService } from './logs/logs.service';
-import { ParseLinks } from "../components/util/parse-links.service";
-
-import { FilterPipe } from "../shared/filter.pipe";
-import { OrderByPipe } from "../shared/order-by.pipe";
-import { TranslatePipe } from '../shared/translate.pipe';
 
 upgradeAdapter.upgradeNg1Provider('$uibModal');
 
@@ -72,19 +60,3 @@ angular
     .directive('<%=jhiPrefix%>Health', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(<%=jhiPrefixCapitalized%>HealthCheckComponent))
     .directive('<%=jhiPrefix%>Logs', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(LogsComponent))
     .factory('<%=jhiPrefixCapitalized%>HealthService', upgradeAdapter.downgradeNg2Provider(<%=jhiPrefixCapitalized%>HealthService));
-
-@NgModule({
-    imports: [BrowserModule, FormsModule],
-    providers: [
-        AuditsService,
-        <%=jhiPrefixCapitalized%>HealthService,
-        LogsService,
-        ParseLinks,
-        {
-            provide: XSRFStrategy,
-            useValue:  new CookieXSRFStrategy('CSRF-TOKEN', 'X-CSRF-TOKEN')
-        }
-    ],
-    declarations: [FilterPipe, OrderByPipe, TranslatePipe]
-})
-export class <%=angularAppName%>AdminModule {}
